requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
httpx
websockets
